Add basic API surface tests for Linux, OSX
I've copied out the existing tests that Windows uses, which log all of the
properties available on various pieces in the library, to separate
versions for Linux and OSX. For those properties that aren't supported on
Linux or OSX, I instead assert that a PlatformNotSupportedException is
thrown. Although this doesn't give us guarantees about the results of
those methods, it does let us know if they have stopped working
altogether.