HttpWebRequest request stream WriteAsync, FlushAsync overrides
This was a prior TODO from PR #2386.

Added overrides for Stream's WriteAsync and FlushAync methods. As the request stream is a memory stream, it'll perform the operations synchronously anyways. Whereas if these methods remain not overridden, it'll default to the base stream's implementation which will queue work items that just call Write and Flush, respectively. So, this improves performance a little bit by not having to queue work items.

Added tests for RequestStream class that is returned via the HttpWebRequest.GetRequstStream* methods.