Add RIDs for OSX 10.11, OS.version, OSfamily-arch
This adds more RIDs as follows:
  OSFamily-arch, eg: unix-x64
  OSX 10.11 El Capitan
  centos.7 synonomous with centos.7.1

experimental RIDs added for consistency with DNX:
  ubuntu.15.04, linuxmint.17, linuxmint.17.1,
  linuxmint.17.2, linuxmint.17.3

Previously we had only created RIDs for runtimes where
needed to pivot assets in our own packages, but this
creates a more complete set to permit 3rd party packages
to define assets more granularly.

I've also moved the runtime.json to the open source.
We still haven't migrated all of the packaging infrastructure
to run from the github repo, but this is a small step in order
to better support collaboration with DNX on this shared
data.

[tfs-changeset: 1544730]