Add support for X509VerificationFlags on Unix
Rather than let OpenSSL stop processing at the first error encountered, implement a callback to collect the errors it reports (and tell it to keep going).

Then, in the Verify step, loop over all errors found, and report on any that are not suppressed by the verification flags.

Note that the use of the verification flags does not remove the information from the ChainElementStatus or ChainStatus arrays, this matches the Windows behavior.