Temporary workaround to get ECDsa tests passing on CentOS
Makes three tests execute conditionally based on whether secp224r1 is enabled.  It's a bit hokey, in that we're using the very mechanism we're intending to test, and further doing string comparison against a known error message, but it's good enough as a temporary workaround, and better than just disabling the tests on all Linux. (Eventually we'll want to look into allowing PlatformSpecific/ActiveIssue to be specialized further.)