Catch exception for unknown charset
When System.Console.dll was first ported to Unix, we didn't have our
globalization stack implemented at all, and Encoding.GetEncoding would
throw a NotSupportedException if the supplied charset wasn't recognized,
so Console was written to catch NotSupportedException to handle the case
of an unknown charset.  Now that our stack is implemented, GetEncoding
throws an ArgumentException as documented on MSDN, and we need to fix
Console to catch the right exception type, otherwise an unknown charset
causes exceptions when attempting to use Console. Further, as arbitrary
encoding providers could throw arbitrary exceptions from GetEncoding, we
need to catch and eat everything.