Changing the order of parameters to EncodeIntoBuffer, which works around #4455 when running on Mono.
When calling stackalloc on Mono, and then immediately passing the size parameter into another method, Mono is zero'ing out the size int. This results in EncodeIntoBuffer throwing an exception because it doesn't think it has any room in the buffer to write the encoded characters.

Switching the order of the parameters appears to work around the issue on Mono, and shouldn't have any negative results on other platforms.

Fix #4455.