Fix terminfo tests to not assume specific files exist
The terminfo tests assume that a specific set of terminfo definition files exist; they do on Ubuntu, CentOS, and OS X, but not all of them do on openSUSE, resulting in test failures.  This commit just makes them optional, bailing out of the test if the file isn't found.  It also cleans up some code to use a theory, now that the bug that was preventing a theory from being used originally has been fixed.