Add HTTP/2 support to HttpClient on Unix
The latest versions of libcurl have HTTP/2 support; this commit enables that to be used from HttpClient via setting the request Version.  Setting the Version will now pass along that version preference to libcurl; if it's set to 2.0, libcurl will use HTTP/2 if supported, otherwise it'll fall back to its default (1.1).  As part of this I also fixed CurlHandler to pass along v1.0 and v1.1 so libcurl can respect the developer's explicit wishes.

(Note that the HTTP/2 light-up is only available if libcurl supports it.  libcurl relies on nghttp2, which needs to be installed, and a version of libcurl needs to be used that's built against it.  To test, I built these locally from source.)