Add signature decoder to System.Reflection.Metadata
Up to now, System.Reflection.Metadata provides low-level access to
ECMA-335 CLI metadata, but only provides signatures as blobs that must
be parsed with direct knowledge of the format as described in the
section II.23.2 of the CLI specification. There is a BlobReader for
reading various elements out of blobs, but it is up to the caller to
read the right things at the right positions.

This was by-design as the library is designed to sit at the lowest level
behind the scenes of higher level API such as Reflection proper, Roslyn,
or CCI.

The challenge with signatures is that they are variable-length and
encode tree structures, and each higher level model that could sit on
top of System.Reflection.Metadata will have different representations
for the trees. We do not want to introduce an API to build a fully
formed tree that then has to be traversed and rewritten to match the
actual use case.

This API is therefore a middle ground between:

 1. Here are the bytes and some helpers. Read the spec to decode.
    (Status quo.)

 2. Yet another high-level metadata API.
    (Out of scope for this layer.)

It works as follows:

 1. The caller chooses an arbitrary representation, TType, for type
    symbols and implements ISignatureTypeProvider<TType> .

 2. Signatures are parsed by recursive descent and the provider is
    called to create new type nodes:

     * Give me the TType that represents this primitive
     * Give me the TType that represents this TypeDefinition
     * Give me the TType that represents an array of this other TType
     * etc.