Address PR feedback.
Updating tests so they run sequentially using the XUnit DisableParallelTests assembly attribute instead of a global lock.

Change all instances of Console.WriteLine to Debug.WriteLine.

Move the DisableParallelTests assembly attribute to a separate file.

Change Debug.WriteLine to use the correct overload in some cases.

Remove ExceptionAssert.cs, it was functionality we used when the tests were running on MSTest, but it exists in XUnit as Assert.Throws.

Clean up the #if false use throughout the tests.

Indicate that USE_ETW is meant for commenting out code until TraceEvent is available.

Change all uses of #if USE_ETW to indicate that they are used to comment out code until TraceEvent is available on CoreCLR.

Remove unnecessary region.

Fix typo in Listeners.cs

Fix typo in EtwListener

Remove Assert.Inconclusive, which is an artifact from when we ran tests on MSTest.

Ran codeformatter on the tests.

Add XUnitAssemblyAttributes.cs to System.Diagnostics.Tracing.Tests.csproj.

Fix Test_WriteEvent_InvalidCalls. The EventSource was not disposed at the end of the test.

Removing a test that was only for the nuget package EventSource.

Remove unecessary TODOs and clean up tests.

Add the ActiveIssue attribute for the tests that are failing on linux/OSX.

Add the ActiveIssue attribute to Test_Write_Fuzzy.

Remove unecessary PlatformID.Linux since PlatformID.AnyUnix already covers it.