Fix SqlParameter Scale and Precision inheritance
The .Net implementation relied on the interfaces to access the
properties on the child class. With the interfaces removed, the child
was not updated to override the member, so getting or setting Scale or
Precision was broken if the object was cast to DbParameter. As noted by
@mgravell in issue dotnet/corefx#1612

This should correct the behavior, but I'm curious why the parent member
are not abstract, instead of stubs.