Support HttpRequestMessage.Version in WinHttpHandler
WinHttpHandler was currently ignoring the `HttpRequestMessage.Version`
property and was always sending out requests using 'HTTP/1.1' which is the
default for winhttp.

This fix now will look at the request message `Version` property and
send out the request using either 1.0 or 1.1.  If neither of these
versions are specified, then the request is sent out using the default
winhttp version (which is currently 1.1 as of Windows 10).

Note: The .NET Framework (Desktop) will throw an exception if neither
1.0 nor 1.1 version is specified.  Currently CoreFx including NETNative,
WinHttpHandler and CurlHandler will simply use a default instead of
throwing. While this is different behavior than Desktop, it is more
desirable in the long run.

There is a separate issue #4870 to add HTTP/2.0 support to WinHttpHandler.

Fixes #4754