Remove redundant OpenSslX509StoreProvider
OpenSslX509StoreProvider was the original stand-in for Windows Store equivalence, and it was mainly tape and bailing wire.

As the Unix X509Certificates implementation expanded eventually there was a CollectionBackedStoreProvider type, which had at least as much functionality.

Since OpenSslX509StoreProvider was one of the last sanctuaries of NotImplementedException in Unix X509Certificates, this type is being removed and the callers migrated over to CollectionBackedStoreProvider.