Make the VerifyCrlCache test resilient to not trusting the root cert.
The default OpenSSL root store on OS X has no certificates in it, causing the VerifyCrlCache test to fail the precondition of a valid chain.  In order to fix that:
* AllowUnknownCertificateAuthority is asserted
* Check that at most one error happened in the precondition test, and that if present it's UntrustedRoot
* All of the error count tests are changed to be relative to the precondition count.

Further resilience could be attained by changing the later tests to accept the RevocationStatusUnknown checks to be allowed at either position, but the list is currently built in leaf-to-root order, so the checks are stable.

Verified on both Ubuntu (trusts the root) and OS X (doesn't).