Improve memory use of HashSet constructed from HashSet.
If a HashSet is constructed from an existing HashSet then use the current
approach if the count is above a particular ratio to the capacity, but if
it is below that ratio (small enough that if created just from those items
and then added to the next resize would still not reach that capacity) use
an optimised version of AddIfNotPresent to add each item.

Fixes #5141

Use count to short-circuit during HashSet ctor

Since we have to examine otherAsHashSet._count anyway, and since
the behaviour of short-circuiting initialising _buckets and
_slots is valid in that case, use that as the test to do so, and
save a little bit of work.

Remove unreachable branch and add test for TrimExcess in ctor.