Specify architecture & OS in native package build recursion
In the open Packages build as part of the managed build
pass, as a platform (architecture) is not set.  Ensure that the
shim packages are built for the correct architectures during
the open build.

In the TFS build we always set platform up front so these
were getting the correct value from the build environment.

Switch to use the OSGroup filtering instead of conditions.