Remove VariableCount
The declared form is never used, and the override only used in an assertion
that can be replaced with a direct call to Variables.Count.

Don't have separate implementation for IEnumerable.GetEnumerator()

Just defering to IEnumerable<Expression>.GetEnumerator() results in one less
hidden type and removes risk of a change causing the two methods to behave
differently.