Only use ECMA key on Windows for System.IO.Compression
There is an issue signing with the ECMA key on non-Windows platforms. Other places in our codebase are not overriding this property if it is explicitly set to 'false' as it is in x-plat builds.

This is also a duplicate property. It is set correctly a few lines below the one I have deleted.