Avoid unnecessary delegate allocations
There are 9 up-front delegate allocations (2 for TextReader and 7 for
TextWriter) that are unnecessary in most cases as most
TextReader/TextWriter subclasses override the base async methods where
such delegates are used. Instead, let the compiler allocate and cache
such delegates lazily, only in the rare cases when the delegates are
actually used.