Improve performance around logging code in System.Net.NameResolution
Same treatment as in https://github.com/dotnet/corefx/pull/5238, but for System.Net.NameResolution.  The call sites to GlobalLog methods often do a lot of allocation and other work, even in release builds and even when logging is disabled; this commit guards those call sites behind checks for whether logging is enabled.