Move failing DNS test to OuterLoop
For a long time now I've been unable to get a clean local build because of
this test.  On the networks I'm usually on, the DNS server serves up a
descriptive page for an invalid address, which causes this test to fail,
as the test is expecting an error.  I'm moving the test to OuterLoop so
that at least normal builds/test runs can pass on such networks.