Workaround [ThreadStatic] limitation in ImmutableList<T>
As a perf improvement awhile back, ImmutableList<T> took a dependency on [ThreadStatic].  This unfortunately breaks cases where ImmutableList<T> is used with a T from a collectible assembly, due to https://github.com/dotnet/coreclr/issues/2191.

Until that limitation is removed, this provides a workaround to restore the ability to use ImmutableList<T> with such types.  Some microbenchmarks around enumeration (which this was originally introduced to help with) show potential throughput regressions up to 10-15%.