Update generation of IsolatedStorage and Loader refs
Both of these packages only carry a single implementation.
IsolatedStorage for netcore50
Loader for dnxcore50

Despite this we still package the reference assembly as
'dotnet'.  As a result, when nuget change the generation
mapping for both netcore50 (5.4 >5.6) and dnxcore50
(5.5 > 5.6) it caused our packaging validation to break.

This is because we have a check in place to make sure
that we don't claim a lower generation than the minimum
supported by all implementations.

I tried changing these reference assemblies to match
the exact TFM of the lib (netcore50 and dnxcore50
respectively) but the validation enforces a dotnet*
reference even if only a single platform is targeted.
I'll look into fixing this.  For now I am updating
the reference projects to unblock the build.

https://github.com/dotnet/corefx/issues/5406

[tfs-changeset: 1564471]