Set SQLClient to dotnet5.4
Fixes https://github.com/dotnet/corefx/issues/5359

SQLClient needs to run on dotnet5.4 but was marking
the x-plat implementation as dotnet5.5.

This was incorrect and prevented folks from building
a 'dotnet' library that would run on Net4.6 and x-plat.

Nothing was preventing SQLClient from claiming
dotnet5.4, it was left behind from prior builds where
some of the System.Net dependencies were not
supported on dotnet5.4.

In doing this I also fixed the windows build to use
the open source rather than closed, and updated
the windows TFM to dotnet5.4 instead of
DNXCore50. Despite using dotnet5.4
we still dummy out the netcore folder in the package
since it won't pass the WACK.

I considered moving the packaging to the open
but currently this depends on SNI.dll which is
not ported.

[tfs-changeset: 1564482]