Fix DirectoryNotFoundException in Process.GetProcessBy*
Most places we access procfs, we catch and eat IOExceptions that result from the process going away while we're trying to read the relevant data from procfs.  But we missed a place where we do such enumeration, causing failures.  This commit just adds a try/catch around one such enumeration of a path to procfs.