Make Console signal handling more robust
In the (probably rare) case where SIGINT and SIGQUIT have been previously set to be ignored (SIG_IGN), we currently end up unhooking our signal handlers, which means that even though we know the app is going to remain alive, we'll stop processing CancelKeyPress events.  This change just makes it so that if we know the signal is going to be ignored, we don't unregister our handlers.