Remove OSGroup=Windows_NT from netcore50/netcore50aot configurations
Originally the goal was to tag the configurations with OSGroup=Windows_NT
if they were only applicable on that OS. However there is nothing particular
to that OS for most of them and they can build and run on an OS that supports
the managed dependencies.

More importantly we want to enable people to build IL only packages on any
OS and not force them to have a separate windows only package for things targeting
netcore50/netcore50aot. As an example System.AppContext recently introduced
an IL only package that can and should build fine on other OS's.