Refactor cross platform Interop types into System.Runtime.InteropServices.PInvoke.dll
SEE https://github.com/dotnet/corefx/pull/5457 for the OSS pull request associated with this change.

Only a subset of the functionality in System.Runtime.InteropServices.dll can be supported cross-platform, so it is necessary to split its contract and partial facade assemblies.

System.Runtime.InteropServices.PInvoke.dll is the new cross-platform contract.

System.Runtime.InteropServices.dll is the "legacy" contract, available on Windows. Referencing allows an app to use the same set of APIs as before, including all of the COM and Windows-specific APIs. It contains a reference to System.Runtime.InteropServices.PInvoke.dll, and type forwards the cross-platform types there.

System.Runtime.InteropServices.Marshal is a special case. Some of its methods are supported cross-platform, and have been moved to a new type - System.Runtime.InteropServices.PInvokeMarshal. Some of the methods' names have been changed, such as AllocCoTaskMem -> AllocateMemory. The forwarding of PInvokeMarshal methods is done using calls in the "implementation" part of the partial facade - see PInvokeMarshal.cs.

GCHandle and GCHandleType have been moved to System.Runtime.dll as part of this general refactoring - they don't have much to do with Interop.

[tfs-changeset: 1567500]