Unix LM\Root X509Store should also read SSL_CERT_FILE
The default resolver in OpenSSL has two providers
1) SSL_CERT_DIR
Every file in this directory has all certificates therein read as a trusted cert.
2) SSL_CERT_FILE
This one extra file is also read, and all certificates therein are trusted certs.

Up until now we've only emulated the SSL_CERT_DIR approach, and that
worked on the Linux distributions tested thus far (Ubuntu 14.04/15.04,
CentOS 7.1, others).

The Homebrew provided OpenSSL for OS X, however, builds the root
store in a single file in the location pointed to by SSL_CERT_FILE.

So, now, read the SSL_CERT_FILE-based content in addition to the
SSL_CERT_DIR content, and things should work better on OS X (and better
match the behavior of an application that would directly be using OpenSSL).