Fix HttpWebRequest to allow content-related headers without content
HttpWebRequest, part of the System.Net.Requests contract, has historically allowed any request header to be sent. However, in CoreFx it is built on top of System.Net.Http. And System.Net.Http has a stricter object model where it requires headers to be added only to specific collections.

The bug was trying to add a content-related header without any content being sent. This worked in .NET Framework (Desktop). While we encourage developers to migrate to the newer System.Net.Http library, we try to maintain app-compat where possible for HttpWebRequest.

The fix is to internally create an empty HttpContent object when calling into System.Net.Http and place the content-related headers into that collection.

Fixes #5565.