Override ReadByte and WriteByte on CryptoStream
Rather than using the base ReadByte and WriteByte methods, which allocate temporary arrays and delegate to Read and Write, respectively, this commit adds overrides of ReadByte and WriteByte to CryptoStream.  In a test that uses these methods repeatedly with AES to write and read 10MB one byte at a time, write throughput was improved by ~5x and read throughput was increased by ~3x.  For both read and write, allocations were reduced by ~10x.